feat: fetch post count data from the new API
