+++ src/services/api/meta-cms/request.ts
@@ -1,0 +1,9 @@
+import { extendWithErrorHandler } from '@/services/api/base-request';
+
+export default extendWithErrorHandler({
+  credentials: 'include',
+  prefix: META_CMS_API,
+  headers: {
+    'Content-Type': 'application/json',
+  },
+});
